Move `deprecationEntries` from `jest-config` to `jest-validate`. Also fix a typo on `jest-runtime`
