+++ .gitignore
@@ -27,3 +27,4 @@ lerna-debug.log
 npm-debug.log
 npm-debug.log*
 yarn-error.log*
+package-lock.json
