+++ packages/jest-validate/package.json
@@ -9,7 +9,6 @@
   "main": "build/index.js",
   "dependencies": {
     "chalk": "^2.0.1",
-    "jest-config": "^22.4.2",
     "jest-get-type": "^22.1.0",
     "leven": "^2.1.0",
     "pretty-format": "^22.4.0"
