+++ packages/jest-validate/src/deprecationEntries.js
